made the count a float so corridors contribute to roomCount but little

+++ src/main/java/teamport/aether/world/generate/feature/components/dungeon/bronze/BaseBronzeRoom.java
@@ -67,7 +67,7 @@ public abstract class BaseBronzeRoom {
     public WorldFeatureComponent decoration;
     public WorldFeatureComponent chest;
     public List<Door> doors;
-    public int roomCount = 1;
+    public float roomCount = 1;
 
     public static BlockPallet ROOM_PALLET = new BlockPallet();
     static {
